refactor(api-compare): unify enumerable-idiom knowledge in shared module
Extract JS_ENUMERABLE_ALIASES into scripts/api-compare/enumerable-idioms.ts
so compare.ts's wide call ratchet and lint-calls.ts's internal call-graph
noise list consume one definition instead of two hand-maintained copies that
drift. compare.ts re-exports the table (test imports unchanged); lint-calls.ts
derives its enumerable-idiom skips from the shared table's keys.

The map contents are moved verbatim, so the narrow and wide call ratchets are
unchanged, and the redundancy guard test still passes.
